misc. updates
