+++ product_odoo_module/README.rst
@@ -46,7 +46,7 @@ Credits
 Authors
 -------
 
-* Quartile Limited
+* Quartile
 
 Maintainers
 -----------
