+++ product_odoo_module/models/res_company.py
@@ -1,0 +1,13 @@
+# Copyright 2025 Quartile (https://www.quartile.co)
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
+from odoo import fields, models
+
+
+class ResCompany(models.Model):
+    _inherit = "res.company"
+
+    price_module_maint = fields.Float(
+        "Module Maintenance Price",
+        help="Standard maintenance cost per LOC (line of code).",
+    )
