+++ product_odoo_module/security/ir.model.access.csv
@@ -1,0 +1,3 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_product_odoo_version_user,product.odoo.version.user,model_product_odoo_version,base.group_user,1,0,0,0
+access_product_odoo_version_manager,product.odoo.version.manager,model_product_odoo_version,base.group_system,1,1,1,1
