+++ product_odoo_module/views/product_odoo_version_views.xml
@@ -1,0 +1,24 @@
+<odoo>
+    <record id="view_product_odoo_version_tree" model="ir.ui.view">
+        <field name="name">product.odoo.version.tree</field>
+        <field name="model">product.odoo.version</field>
+        <field name="arch" type="xml">
+            <tree editable="bottom">
+                <field name="name" />
+                <field name="active" />
+            </tree>
+        </field>
+    </record>
+    <record id="menu_action_odoo_version" model="ir.actions.act_window">
+        <field name="name">Odoo Versions</field>
+        <field name="res_model">product.odoo.version</field>
+        <field name="view_mode">tree</field>
+    </record>
+    <menuitem
+        id="menu_odoo_version"
+        name="Odoo Versions"
+        parent="product_odoo_module.menu_odoo_module_in_config_sales"
+        action="menu_action_odoo_version"
+        sequence="10"
+    />
+</odoo>
