+++ product_odoo_module/views/product_template_views.xml
@@ -1,8 +1,8 @@
 <odoo>
-    <record id="product_template_form_view_inherit" model="ir.ui.view">
-        <field name="name">product.template.form.inherit</field>
+    <record id="product_template_form_view" model="ir.ui.view">
+        <field name="name">product.template.common.form</field>
         <field name="model">product.template</field>
-        <field name="inherit_id" ref="product.product_template_only_form_view" />
+        <field name="inherit_id" ref="product.product_template_form_view" />
         <field name="arch" type="xml">
             <xpath expr="//div[@name='options']" position="inside">
                 <span class="d-inline-block">
@@ -16,9 +16,30 @@
                     attrs="{'invisible': [('is_odoo_module', '=', False)]}"
                 >
                     <group>
-                        <field name="technical_name" />
+                        <field name="license_type" />
+                        <field name="odoo_version_id" />
+                        <field name="complexity" />
+                        <field name="popularity_factor" />
+                        <field name="loc" />
+                        <field name="amount" />
+                        <field name="amount_incl_mig" />
                     </group>
                 </page>
+            </xpath>
+        </field>
+    </record>
+    <record id="product_template_search_view" model="ir.ui.view">
+        <field name="name">product.template.search</field>
+        <field name="model">product.template</field>
+        <field name="inherit_id" ref="product.product_template_search_view" />
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='attribute_line_ids']" position="before">
+                <filter
+                    string="Odoo Module"
+                    name="filter_odoo_module"
+                    domain="[('is_odoo_module','=',True)]"
+                />
+                <separator />
             </xpath>
         </field>
     </record>
