+++ product_odoo_module/views/res_company_views.xml
@@ -1,0 +1,14 @@
+<odoo>
+    <record id="view_company_form" model="ir.ui.view">
+        <field name="name">res.company.form</field>
+        <field name="model">res.company</field>
+        <field name="inherit_id" ref="base.view_company_form" />
+        <field name="arch" type="xml">
+            <xpath expr="//group[@name='social_media']" position="after">
+                <group string="Odoo Module Attributes" name="odoo_module_attributes">
+                    <field name="price_module_maint" widget="monetary" />
+                </group>
+            </xpath>
+        </field>
+    </record>
+</odoo>
